bugfix: Insert imports in the correct place in ScalaCLI scripts
Previously, we would expect a file to be a ScalaCLI script if it ended in `.sc`, but that seems to have changed and now it ends in `.sc.scala` similar to ammonite. Now, we adjusted the extension to account for that difference.

Fixes https://github.com/scalameta/metals/issues/4939
