+++ project/TestGroups.scala
@@ -103,7 +103,7 @@ object TestGroups {
       "tests.codeactions.ConvertToNamedArgumentsLspSuite",
       "tests.testProvider.TestSuitesProviderSuite", "tests.MillVersionSuite",
       "tests.FingerprintsLspSuite", "tests.JdkVersionSuite",
-      "tests.codeactions.ActionableDiagnosticsSuite",
+      "tests.codeactions.ScalaCliActionsSuite",
       "tests.codeactions.MillifyDependencyLspSuite",
       "tests.RenameFilesLspSuite", "tests.codeactions.ExtractMethodLspSuite",
       "tests.SemanticTokensExpectSuite",
